Multiple submeshes per part.
