+++ spine-unity/Assets/spine-unity/Mesh Generation/Arrays/ArraysBuffers.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 7492a8ff24ea1884f92307ab1121e451
-timeCreated: 1458046464
+guid: 043e864df13214e4989d29c17a863b08
+timeCreated: 1458056200
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
