+++ spine-unity/Assets/spine-unity/Mesh Generation/Arrays/ArraysSimpleMeshGenerator.cs
@@ -32,7 +32,7 @@
 using UnityEngine;
 using System.Collections;
 
-namespace Spine.Unity {
+namespace Spine.Unity.MeshGeneration {
 	public class ArraysSimpleMeshGenerator : ISimpleMeshGenerator {
 		#region Settings
 		protected float scale = 1f;
@@ -42,6 +42,8 @@ namespace Spine.Unity {
 		}
 
 		public bool renderMeshes = true;
+
+		public bool premultiplyVertexColors = true;
 		#endregion
 
 		#region Buffers
@@ -112,14 +114,11 @@ namespace Spine.Unity {
 
 			// Step 3 : Push vertices to arrays
 			//
-			const float z = 0;
+			const float zSpacing = 0;
 			const float zFauxHalfThickness = 0.01f;	// Somehow needs this thickness for bounds to work properly in some cases (eg, Unity UI clipping)
 			float[] tempVertices = this.tempVertices;
 			Vector2[] uvs = this.uvs;
 			Color32[] colors = this.colors;
-			int vertexIndex = 0;
-			Color32 color;
-			float a = skeleton.a * 255, r = skeleton.r, g = skeleton.g, b = skeleton.b;
 
 			Vector3 meshBoundsMin;
 			Vector3 meshBoundsMax;
@@ -134,116 +133,23 @@ namespace Spine.Unity {
 				meshBoundsMin.z = -zFauxHalfThickness;
 				meshBoundsMax.z = zFauxHalfThickness;
 
-				int i = 0;
-				do {
-					Slot slot = drawOrderItems[i];
-					Attachment attachment = slot.attachment;
-					var regionAttachment = attachment as RegionAttachment;
-					if (regionAttachment != null) {
-						regionAttachment.ComputeWorldVertices(slot.bone, tempVertices);
+				int vertexIndex = 0;
+				ArraysBuffers.Fill(skeleton, 0, drawOrderCount, zSpacing, this.premultiplyVertexColors, vertices, uvs, colors, ref vertexIndex, ref tempVertices, ref meshBoundsMin, ref meshBoundsMax);
+				this.tempVertices = tempVertices;
 
-						float x1 = tempVertices[RegionAttachment.X1], y1 = tempVertices[RegionAttachment.Y1];
-						float x2 = tempVertices[RegionAttachment.X2], y2 = tempVertices[RegionAttachment.Y2];
-						float x3 = tempVertices[RegionAttachment.X3], y3 = tempVertices[RegionAttachment.Y3];
-						float x4 = tempVertices[RegionAttachment.X4], y4 = tempVertices[RegionAttachment.Y4];
-						vertices[vertexIndex].x = x1 * scale; vertices[vertexIndex].y = y1 * scale;	vertices[vertexIndex].z = z;
-						vertices[vertexIndex + 1].x = x4 * scale; vertices[vertexIndex + 1].y = y4 * scale; vertices[vertexIndex + 1].z = z;
-						vertices[vertexIndex + 2].x = x2 * scale; vertices[vertexIndex + 2].y = y2 * scale; vertices[vertexIndex + 2].z = z;
-						vertices[vertexIndex + 3].x = x3 * scale; vertices[vertexIndex + 3].y = y3 * scale; vertices[vertexIndex + 3].z = z;
+				// Apply scale to vertices
+				for (int i = 0; i < totalVertexCount; i++) {
+					var v = vertices[i];
+					v.x *= scale;
+					v.y *= scale;
+					vertices[i] = v;
+				}
 
-						color.a = (byte)(a * slot.a * regionAttachment.a);
-						color.r = (byte)(r * slot.r * regionAttachment.r * color.a);
-						color.g = (byte)(g * slot.g * regionAttachment.g * color.a);
-						color.b = (byte)(b * slot.b * regionAttachment.b * color.a);
-						if (slot.data.blendMode == BlendMode.additive) color.a = 0;
-						colors[vertexIndex] = color; colors[vertexIndex + 1] = color; colors[vertexIndex + 2] = color; colors[vertexIndex + 3] = color;
+				meshBoundsMax.x *= scale;
+				meshBoundsMax.y *= scale;
+				meshBoundsMin.x *= scale;
+				meshBoundsMax.y *= scale;
 
-						float[] regionUVs = regionAttachment.uvs;
-						uvs[vertexIndex].x = regionUVs[RegionAttachment.X1]; uvs[vertexIndex].y = regionUVs[RegionAttachment.Y1];
-						uvs[vertexIndex + 1].x = regionUVs[RegionAttachment.X4]; uvs[vertexIndex + 1].y = regionUVs[RegionAttachment.Y4];
-						uvs[vertexIndex + 2].x = regionUVs[RegionAttachment.X2]; uvs[vertexIndex + 2].y = regionUVs[RegionAttachment.Y2];
-						uvs[vertexIndex + 3].x = regionUVs[RegionAttachment.X3]; uvs[vertexIndex + 3].y = regionUVs[RegionAttachment.Y3];
-
-						// Calculate min/max X
-						if (x1 < meshBoundsMin.x) meshBoundsMin.x = x1;
-						else if (x1 > meshBoundsMax.x) meshBoundsMax.x = x1;
-						if (x2 < meshBoundsMin.x) meshBoundsMin.x = x2;
-						else if (x2 > meshBoundsMax.x) meshBoundsMax.x = x2;
-						if (x3 < meshBoundsMin.x) meshBoundsMin.x = x3;
-						else if (x3 > meshBoundsMax.x) meshBoundsMax.x = x3;
-						if (x4 < meshBoundsMin.x) meshBoundsMin.x = x4;
-						else if (x4 > meshBoundsMax.x) meshBoundsMax.x = x4;
-
-						// Calculate min/max Y
-						if (y1 < meshBoundsMin.y) meshBoundsMin.y = y1;
-						else if (y1 > meshBoundsMax.y) meshBoundsMax.y = y1;
-						if (y2 < meshBoundsMin.y) meshBoundsMin.y = y2;
-						else if (y2 > meshBoundsMax.y) meshBoundsMax.y = y2;
-						if (y3 < meshBoundsMin.y) meshBoundsMin.y = y3;
-						else if (y3 > meshBoundsMax.y) meshBoundsMax.y = y3;
-						if (y4 < meshBoundsMin.y) meshBoundsMin.y = y4;
-						else if (y4 > meshBoundsMax.y) meshBoundsMax.y = y4;
-
-						vertexIndex += 4;
-					} else {
-						if (!renderMeshes) continue;
-						var meshAttachment = attachment as MeshAttachment;
-						if (meshAttachment != null) {
-							int meshVertexCount = meshAttachment.vertices.Length;
-							if (tempVertices.Length < meshVertexCount)
-								this.tempVertices = tempVertices = new float[meshVertexCount];
-							meshAttachment.ComputeWorldVertices(slot, tempVertices);
-
-							color.a = (byte)(a * slot.a * meshAttachment.a);
-							color.r = (byte)(r * slot.r * meshAttachment.r * color.a);
-							color.g = (byte)(g * slot.g * meshAttachment.g * color.a);
-							color.b = (byte)(b * slot.b * meshAttachment.b * color.a);
-							if (slot.data.blendMode == BlendMode.additive) color.a = 0;
-
-							float[] meshUVs = meshAttachment.uvs;
-							for (int ii = 0; ii < meshVertexCount; ii += 2, vertexIndex++) {
-								float x = tempVertices[ii], y = tempVertices[ii + 1];
-								vertices[vertexIndex].x = x * scale; vertices[vertexIndex].y = y * scale; vertices[vertexIndex].z = z;
-								colors[vertexIndex] = color;
-								uvs[vertexIndex].x = meshUVs[ii]; uvs[vertexIndex].y = meshUVs[ii + 1];
-
-								if (x < meshBoundsMin.x) meshBoundsMin.x = x;
-								else if (x > meshBoundsMax.x) meshBoundsMax.x = x;
-
-								if (y < meshBoundsMin.y) meshBoundsMin.y = y;
-								else if (y > meshBoundsMax.y) meshBoundsMax.y = y;
-							}
-						} else {
-							var skinnedMeshAttachment = attachment as WeightedMeshAttachment;
-							if (skinnedMeshAttachment != null) {
-								int meshVertexCount = skinnedMeshAttachment.uvs.Length;
-								if (tempVertices.Length < meshVertexCount)
-									this.tempVertices = tempVertices = new float[meshVertexCount];
-								skinnedMeshAttachment.ComputeWorldVertices(slot, tempVertices);
-
-								color.a = (byte)(a * slot.a * skinnedMeshAttachment.a);
-								color.r = (byte)(r * slot.r * skinnedMeshAttachment.r * color.a);
-								color.g = (byte)(g * slot.g * skinnedMeshAttachment.g * color.a);
-								color.b = (byte)(b * slot.b * skinnedMeshAttachment.b * color.a);
-								if (slot.data.blendMode == BlendMode.additive) color.a = 0;
-
-								float[] meshUVs = skinnedMeshAttachment.uvs;
-								for (int ii = 0; ii < meshVertexCount; ii += 2, vertexIndex++) {
-									float x = tempVertices[ii], y = tempVertices[ii + 1];
-									vertices[vertexIndex].x = x * scale; vertices[vertexIndex].y = y * scale; vertices[vertexIndex].z = z;
-									colors[vertexIndex] = color;
-									uvs[vertexIndex].x = meshUVs[ii]; uvs[vertexIndex].y = meshUVs[ii + 1];
-
-									if (x < meshBoundsMin.x) meshBoundsMin.x = x;
-									else if (x > meshBoundsMax.x) meshBoundsMax.x = x;
-
-									if (y < meshBoundsMin.y) meshBoundsMin.y = y;
-									else if (y > meshBoundsMax.y) meshBoundsMax.y = y;
-								}
-							}
-						}
-					}
-				} while (++i < drawOrderCount);
 			}
