+++ spine-unity/Assets/spine-unity/Mesh Generation/Arrays/ArraysSubmeshSetMeshGenerator.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: d8742ee1d54c9954d95c9e6508806ac2
-timeCreated: 1457405832
+guid: 11ba077d7d984814db31d054192be532
+timeCreated: 1458047178
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
