+++ spine-unity/Assets/spine-unity/Mesh Generation/Arrays/SubmeshTriangleBuffer.cs
@@ -1,0 +1,14 @@
+﻿using UnityEngine;
+using System.Collections;
+
+namespace Spine.Unity.MeshGeneration {
+	public class SubmeshTriangleBuffer {
+		public int[] triangles;
+		public int triangleCount;
+
+		public SubmeshTriangleBuffer (int triangleCount) {
+			triangles = new int[triangleCount];
+			this.triangleCount = triangleCount;
+		}
+	}
+}
