+++ spine-unity/Assets/spine-unity/Mesh Generation/Arrays/SubmeshTriangleBuffer.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 15418e02758268b488e9ab86f805330c
-timeCreated: 1457407060
+guid: 4b4bb4dc3775d6548b9405f681601abb
+timeCreated: 1458056219
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
