+++ spine-unity/Assets/spine-unity/Modules/SkeletonRenderSeparator/Editor/SkeletonRenderPartInspector.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 7492a8ff24ea1884f92307ab1121e451
-timeCreated: 1458046464
+guid: 30e43037bf4433645ad70266f34c1c8b
+timeCreated: 1458051036
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
