+++ spine-unity/Assets/spine-unity/Modules/SkeletonRenderSeparator/Editor/SkeletonRenderSeparatorInspector.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 5273e5c60bd84a44db5a4e8f3252eef0
-timeCreated: 1457396525
+guid: d2a5062cfe5dd4344831cda4723128af
+timeCreated: 1458067064
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
