+++ spine-unity/Assets/spine-unity/Modules/SkeletonRenderSeparator/SkeletonRenderPart.cs
@@ -1,13 +1,12 @@
 ﻿using UnityEngine;
 using System.Collections;
-
-using UnityEngine.Assertions;
+using Spine.Unity.MeshGeneration;
 
 namespace Spine.Unity {
 	[RequireComponent(typeof(MeshRenderer), typeof(MeshFilter))]
-	public class SubmeshRenderer : MonoBehaviour {		
-		ISingleSubmeshGenerator meshGenerator;
-		public ISingleSubmeshGenerator MeshGenerator {
+	public class SkeletonRenderPart : MonoBehaviour {		
+		ISubmeshSetMeshGenerator meshGenerator;
+		public ISubmeshSetMeshGenerator MeshGenerator {
 			get {
 				LazyIntialize();
 				return meshGenerator;
@@ -32,7 +31,7 @@ namespace Spine.Unity {
 
 		void LazyIntialize () {
 			if (meshGenerator != null) return;
-			meshGenerator = new ArraysSingleSubmeshGenerator(); // swap this out with your custom ISingleSubmeshGenerator code.
+			meshGenerator = new ArraysSubmeshSetMeshGenerator();
 			meshFilter = GetComponent<MeshFilter>();
 			meshRenderer = GetComponent<MeshRenderer>();
 		}
@@ -41,10 +40,11 @@ namespace Spine.Unity {
 			meshFilter.sharedMesh = null;
 		}
 
-		public void RenderSubmesh (SubmeshInstruction instruction) {
+		public void RenderSubmesh (ExposedList<SubmeshInstruction> instructions, int startSubmesh, int endSubmesh) {
 			LazyIntialize();
-			meshRenderer.sharedMaterial = instruction.material;
-			meshFilter.sharedMesh = meshGenerator.GenerateMesh(instruction);
+			MeshAndMaterials m = meshGenerator.GenerateMesh(instructions, startSubmesh, endSubmesh);
+			meshFilter.sharedMesh = m.mesh;
+			meshRenderer.sharedMaterials = m.materials;
 		}
 
 		public void SetPropertyBlock (MaterialPropertyBlock block) {
@@ -52,10 +52,10 @@ namespace Spine.Unity {
 			meshRenderer.SetPropertyBlock(block);
 		}
 
-		public static SubmeshRenderer NewSubmeshRendererGameObject (Transform parent, string name) {
+		public static SkeletonRenderPart NewSubmeshRendererGameObject (Transform parent, string name) {
 			var go = new GameObject(name, typeof(MeshFilter), typeof(MeshRenderer));
 			go.transform.SetParent(parent, false);
-			var returnComponent = go.AddComponent<SubmeshRenderer>();
+			var returnComponent = go.AddComponent<SkeletonRenderPart>();
 
 			return returnComponent;
 		}
