+++ spine-unity/Assets/spine-unity/Modules/SkeletonRenderSeparator/SkeletonRenderPart.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 7492a8ff24ea1884f92307ab1121e451
-timeCreated: 1458046464
+guid: 1c0b968d1e7333b499e347acb644f1c1
+timeCreated: 1458045480
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
