--- spine-unity/Assets/spine-unity/Modules/SubmeshRenderer.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 792af191c395561459cc0cc03a517ae8
-folderAsset: yes
-timeCreated: 1457401769
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
