--- spine-unity/Assets/spine-unity/Modules/SubmeshRenderer/Editor.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 5eec1664658bb4344909619707652674
-folderAsset: yes
-timeCreated: 1457405608
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
